% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readAgilent.R
\name{combinePeakList.AgilentExport}
\alias{combinePeakList.AgilentExport}
\title{combinePeakList.AgilentExport}
\usage{
combinePeakList.AgilentExport(peakList, removeNAColumns = TRUE)
}
\arguments{
\item{peakList}{list that contains two elements: info (data.frame) and data
(list of data.frame's). Will usually be the result of either \code{readPeaklist.AgilentExport.memory}
or \code{readPeaklist.AgilentExport}}

\item{removeNAColumns}{logical vector. If TRUE (default) then all columns of
the resulting data.frame that only contain NA's are removed. Default: TRUE}
}
\value{
a data.frame (or NA)
}
\description{
combines the results from \code{readPeaklist.AgilentExport.memory}
 or \code{readPeaklist.AgilentExport} into a single data.frame.
}
\examples{
demoFile <- fs::path_package("extdata", "raw/WorklistData-0005_MS.csv", package = "AgilentDataV2")
result <- readPeaklist.AgilentExport(demoFile, startsString = "WorklistData-0005")()
combinePeakList.AgilentExport(result) |> head(10)

}
