% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_saveSignals.R
\name{completeTable2Signals}
\alias{completeTable2Signals}
\title{completeTable2Signals}
\usage{
completeTable2Signals(theTable)
}
\arguments{
\item{theTable}{data.frame, should be the result from the funciton getMzs}
}
\value{
a data.frame. If theTable is NA, then the function will return an
 empty data.frame
}
\description{
function that converts the result from the function getMzs
 (data.frame, which is more or less end result) to data.frame that can be
 exported to a .csv file ready to be imported to the correct GxP worksheet in
 GxP Signals Notebooks
}
\examples{
demoSpecFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005.mzdata.xml",
  package = "AgilentDataV2"
)
demoPeaksFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005_MS.csv",
  package = "AgilentDataV2"
)
demoTargetsFile <- fs::path_package(
  "extdata",
  "targets/targets001.csv",
  package = "AgilentDataV2"
)
result <- getData.All(
  sampleName = "WKL-0005",
  peptideTargetsFile = demoTargetsFile,
  mzDataFile = demoSpecFile,
  msPeakFile = demoPeaksFile,
)
thePeaks <- getRRtArea(theData = result[[1]])
getMzs(theData = result[[1]],
       thePeaks = thePeaks,
       mzNr = 1)
result <- getMzs(theData = result[[1]],
                 thePeaks = thePeaks,
                 mzNr = NA)
completeTable2Signals(result)
completeTable2Signals(NA)
}
