% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{countProcessingData}
\alias{countProcessingData}
\title{countProcessingData}
\usage{
countProcessingData(theFilenames, includeIncomplete = FALSE)
}
\arguments{
\item{theFilenames}{two element named list of character vectors (filenames
that may match)}

\item{includeIncomplete}{logical vector. Default is FALSE, then only complete
processing (matched) files are counter. If TRUE then all (matched and
unmatched) are counter}
}
\value{
integer vector
}
\description{
Function that returns the number of (complete or incomplete) processing files
}
\examples{
processingFiles <- list(
  MSPeakFiles = c("./20250702/WorklistData-0003_MS.csv",
                  "./20250702/WorklistData-0005_MS.csv"),
  Spectra = c("./20250702/WorklistData-0003.mzdata.xml",
              "./20250702/WorklistData-0005.mzdata.xml")
)
processingFilesError1 <- list(
  MSPeakFiles = c("./20250702/WorklistData-0003_MS.csv"),
  Spectra = c("./20250702/20250702/WorklistData-0003.mzdata.xml",
              "./20250702/WorklistData-0005.mzdata.xml")
)
countProcessingData(processingFiles)
countProcessingData(processingFilesError1)
countProcessingData(processingFilesError1, includeIncomplete = TRUE)
}
