% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{emptyFiles}
\alias{emptyFiles}
\title{emptyFiles}
\usage{
emptyFiles(theFiles = NA)
}
\arguments{
\item{theFiles}{the variable to be tested}
}
\value{
logical vector
}
\description{
Helper function that tests if a processingFiles vector is empty
}
\note{
This function tests if the parameter theFiles is
 \code{link[base]{identical}} to the result of the processingFilesEmpty
 function. If it is identical, the function will return FALSE. All other
 possibilities result in TRUE. The structure of the theFiles parameter is
 NOT checked
}
\examples{
emptyFiles()
processingFiles <- list(
  MSPeakFiles = c("WorklistData-0003_MS.csv",
                  "WorklistData-0005_MS.csv"),
  Spectra = c("WorklistData-0003.mzdata.xml",
              "WorklistData-0005.mzdata.xml")
)
emptyFiles(theFiles = processingFiles)
}
