% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tools.R
\name{fieldsPresent}
\alias{fieldsPresent}
\title{fieldsPresent}
\usage{
fieldsPresent(theList, fields, allowNA = FALSE)
}
\arguments{
\item{theList}{list object with named elements}

\item{fields}{character vector of the names which are to be tested for being
present}

\item{allowNA}{if FALSE (default) then any element with a name in 'fields'
present which is NA will be counted as not present. If TRUE, then this is
ignored}
}
\value{
a logical vector
}
\description{
function that checks if one or more elements with a defined name
 are present in a list. It also can check if the fields (if present) are NA
}
\examples{
testList <- list(
 mtcars = mtcars,
 iris = iris,
 test = NA
)
fieldsPresent(theList = testList, fields = c("mtcars", "iris"))
fieldsPresent(theList = testList, fields = c("mtcars", "airquality"))
fieldsPresent(theList = testList, fields = c("mtcars", "test"))
fieldsPresent(theList = testList, fields = c("mtcars", "test"), allowNA = TRUE)
}
