% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{gatherAllCompleteFiles}
\alias{gatherAllCompleteFiles}
\title{gatherAllCompleteFiles}
\usage{
gatherAllCompleteFiles(theFiles)
}
\arguments{
\item{theFiles}{two element named list of character vectors (filenames to be
matched)}
}
\value{
#' @returns a two element list of the filenames in 'theFiles'
 argument, which have been matched. Filenames which have not been matched are
 excluded
}
\description{
Function that extracts the filenames which have matches (and thus have both
 MSPeakFiles and Spectra). See examples.
}
\examples{
processingFiles <- list(
  MSPeakFiles = c("./20250702/WorklistData-0003_MS.csv",
                  "./20250702/WorklistData-0005_MS.csv"),
  Spectra = c("./20250702/WorklistData-0003.mzdata.xml",
              "./20250702/WorklistData-0005.mzdata.xml")
)
processingFilesError1 <- list(
  MSPeakFiles = c("./20250702/WorklistData-0003_MS.csv"),
  Spectra = c("./20250702/WorklistData-0003.mzdata.xml",
              "./20250702/WorklistData-0005.mzdata.xml")
)
processingFilesError2 <- list(
  MSPeakFiles = c("./20250702/WorklistData-0003_MS.csv",
                  "./20250702/WorklistData-0005_MS.csv"),
  Spectra = c("./20250702/WorklistData-0005.mzdata.xml")
)
gatherAllCompleteFiles(processingFiles)
gatherAllCompleteFiles(processingFilesError1)
gatherAllCompleteFiles(processingFilesError2)
}
