% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{gatherRelatedFiles}
\alias{gatherRelatedFiles}
\title{gatherRelatedFiles}
\usage{
gatherRelatedFiles(theFiles, sample)
}
\arguments{
\item{theFiles}{two element named list of character vectors (filenames to be
matched)}

\item{sample}{(part of) the name of the file to be matched}
}
\value{
a list of the filenames in 'theFiles' argument, which have the
 'sample' argument
}
\description{
function that attempts to get the filenames which 'belong' together, so that
 the correct MSPeakFile is matched with the correct Spectra file
}
\examples{
processingFiles <- list(
  MSPeakFiles = c("./20250702/WorklistData-0003_MS.csv",
                  "./20250702/WorklistData-0005_MS.csv"),
  Spectra = c("./20250702/WorklistData-0003.mzdata.xml",
              "./20250702/WorklistData-0005.mzdata.xml")
)
gatherRelatedFiles(theFiles = processingFiles,
                   sample = "WorklistData-0005")
gatherRelatedFiles(theFiles = processingFiles,
                   sample = "WorklistData-0003")
}
