% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read.R
\name{getData.All}
\alias{getData.All}
\title{getData.All}
\usage{
getData.All(
  sampleName = "Data",
  listData = list(),
  sendMessage = print,
  updateProgressor = NA,
  peptideTargetsFile = NA,
  getTargets = getData.Targets,
  mzDataFile,
  getSpectra = getData.Spectra,
  msPeakFile,
  getMSPeakFiles = getData.MSPeakFiles,
  useSmooth = FALSE
)
}
\arguments{
\item{sampleName}{character vector. Name assigned to the resulting
list-element in the listData list}

\item{listData}{list object, to which the results of this function are added
(as a list element)}

\item{sendMessage}{function to print progress messages. Default = print. To
prevent messages from being printed, set to e.g.
sendMessage = function(...){}}

\item{updateProgressor}{argument to pass on a shiny Progress object. See also
\code{link[shiny]{Progress}}. Progress object must be initialized. Default
is NA (no shiny Progress)}

\item{peptideTargetsFile}{character vector or data.frame/list with an column/element
called 'file': filename + path to a csv file that contains the peptide target
info. For this function the exact columns and contents are not important. It
needs to have headers and separation character is ';'}

\item{getTargets}{function that defines how to read the targets file. Default
is \code{link{getData.Targets}}}

\item{mzDataFile}{mzDataFile filename (including path) of the spectra data}

\item{getSpectra}{function that defines how to read the spectra data
(mzdata.xml file). Default is \code{link{getData.Spectra}}}

\item{msPeakFile}{filename (including path) of the peak info file}

\item{getMSPeakFiles}{function that defines how to read the peak info files.
Default is \code{link{getData.MSPeakFiles}}}

\item{useSmooth}{logical vector that defines whether smoothed data is removed
(FALSE, default) or if raw data is removed (TRUE). If NA, then no filtering
takes place.}
}
\value{
listData with added a list element that contains the data read
}
\description{
Reads all the specified spectra, peak files & targets. Prints messages to
 the R console to allow one to follow progress. For use in shiny
 applications, there are options to use progress reporters. See
 \code{link[shiny]{Progress}} for more information
}
\examples{
demoSpecFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005.mzdata.xml",
  package = "AgilentDataV2"
)
demoPeaksFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005_MS.csv",
  package = "AgilentDataV2"
)
demoTargetsFile <- fs::path_package(
  "extdata",
  "targets/targets001.csv",
  package = "AgilentDataV2"
)
result <- getData.All(
  sampleName = "WKL-0005",
  peptideTargetsFile = demoTargetsFile,
  mzDataFile = demoSpecFile,
  msPeakFile = demoPeaksFile,
)
result$`WKL-0005`$targets
result$`WKL-0005`$sampleName
result[[1]]$spectra
result[[1]]$peakList
}
