% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readAgilent.R
\name{getData.MSPeakFiles}
\alias{getData.MSPeakFiles}
\title{getData.MSPeakFiles}
\usage{
getData.MSPeakFiles(
  filename = NA,
  startsString =
    paste0(stringr::str_remove(tools::file_path_sans_ext(basename(filename)), pattern =
    "_.*"), ".d"),
  useSmooth = FALSE,
  removeNAColumns = TRUE
)
}
\arguments{
\item{filename}{character vector: filename (including path) of the peak info
file}

\item{startsString}{character vector that defines which lines are the start
of a peaklist lines in the file}

\item{useSmooth}{logical vector that defines whether smoothed data is removed
(FALSE) or if raw data is removed (TRUE). If NA, then nothing, then no data
is removed. Note: When chromatograms are smoothed, we end up with two
chromatograms per trace: the raw & the smoothed one. If smoothing is done
prior to peak detection/integration, then there will also be two peak tables
per trace. This can lead to confusion in code where data is combined. This
option removes either the smoothed data or the raw data. Default is FALSE}

\item{removeNAColumns}{logical vector. If TRUE (default) then all columns of
the resulting info data.frame that only contain NA's are removed.
Default: TRUE}
}
\value{
list or NA (in case of fail)
}
\description{
wrapper function that reads the peak info file (integration data) and combines
 the info & data into a single data.frame
}
\examples{
demoFile <- fs::path_package("extdata", "raw/WorklistData-0005_MS.csv", package = "AgilentDataV2")
getData.MSPeakFiles(filename = demoFile, startsString = "WorklistData-0005") |> head(10)

}
