% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectraAgilent.R
\name{getData.Spectra}
\alias{getData.Spectra}
\title{getData.Spectra}
\usage{
getData.Spectra(filename = NA)
}
\arguments{
\item{filename}{character vector: filename (including path) of the spectra
data}
}
\value{
an mzData object or NA
}
\description{
reads the spectra data file (.mzdata.xml). Uses readFile internally to
 prevent stops due to e.g. files not existing
}
\examples{
demoFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005.mzdata.xml",
  package = "AgilentDataV2"
)
spectrData <- getData.Spectra(filename = demoFile)
spectrData
spectrData$info(centroided = TRUE)
plot(
 spectrData$spectrum(number = 3),
 type = "h",
 xlab = "m/z", ylab = 'Intensity'
)
plot(
 spectrData$spectrum(number = 4),
 type = "h",
 xlab = "m/z", ylab = 'Intensity'
)

}
