% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read.R
\name{getData.Targets}
\alias{getData.Targets}
\title{getData.Targets}
\usage{
getData.Targets(peptideTargetsFile = NA)
}
\arguments{
\item{peptideTargetsFile}{character vector or data.frame/list with an column/element
called 'file': filename + path to a csv file that contains the peptide target
info. For this function the exact columns and contents are not important. It
needs to have headers and separation character is ';'}
}
\value{
a data.frame or NA (in case of failure to read)
}
\description{
Function that reads the peptide targets file
}
\examples{
demoFile <- fs::path_package(
 "extdata",
 "targets/targets001.csv",
 package = "AgilentDataV2"
)
demoFile
getData.Targets(peptideTargetsFile = demoFile)
}
