% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peakList.R
\name{getMzs}
\alias{getMzs}
\title{getMzs}
\usage{
getMzs(
  theData,
  thePeaks = NA,
  mzNr = 1,
  setSingleRt = 0.05,
  limitPpm = c(30, 30),
  peakIntensityLimit = 10
)
}
\arguments{
\item{theData}{usually the result from \code{link{getData.All}}}

\item{thePeaks}{usually the result from \code{link{getRtArea}} or
\code{link{getRRtArea}}}

\item{mzNr}{rownumber in the targets data.frame to search the peak info for.
If NA then all rows will be searched individually (may take some time) and
returned in data.frame. If integer value then it will search only for the
'component' in that specific row of the targets data.frame. If a component
is not found, them the peak info will be all NA's}

\item{setSingleRt}{numeric vector: retention time in minutes to subtract and
add to the retention time of single spectrum spectra. Spectra are usually a
sum or average of a number of spectra close together. The lowest Rt spectrum
provides the start (Rt), the highest Rt spectrum provides the end (Rt). In
case a single spectrum provides the data then start = Rt - setSingleRt and
end = Rt + setSingleRt. This is to accomodate other code where both start
and end are required}

\item{limitPpm}{how much the m/z in a spectrum is allowed to deviate. All
m/z's inside this range (m/z - ppm, m/z + ppm) are used to generate the
foundmz}

\item{peakIntensityLimit}{the minimum intensity an m/z peak in a spectrum
must have to used (elimination of noise).}
}
\value{
thePeaks data.frame with extra column for the m/z (foundmz) (or NA)
}
\description{
Attemps to get the m/z info for all found peaks in the data
}
\examples{
demoSpecFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005.mzdata.xml",
  package = "AgilentDataV2"
)
demoPeaksFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005_MS.csv",
  package = "AgilentDataV2"
)
demoTargetsFile <- fs::path_package(
  "extdata",
  "targets/targets001.csv",
  package = "AgilentDataV2"
)
result <- getData.All(
  sampleName = "WKL-0005",
  peptideTargetsFile = demoTargetsFile,
  mzDataFile = demoSpecFile,
  msPeakFile = demoPeaksFile,
)
thePeaks <- getRRtArea(theData = result[[1]])
getMzs(theData = result[[1]],
       thePeaks = thePeaks,
       mzNr = 1)
\dontrun{
getMzs(theData = result[[1]],
       thePeaks = thePeaks,
       mzNr = NA)
}
}
