% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peakList.R
\name{getRRtArea}
\alias{getRRtArea}
\title{getRRtArea}
\usage{
getRRtArea(
  theData,
  closeRts = list(left = 0.05, right = 0.05, closest = TRUE, peak = FALSE),
  refcloseRts = list(left = 0.5, right = 0.5, closest = TRUE, peak = FALSE),
  fixedRrt = NA,
  includeMissing = TRUE,
  mzNr = NA
)
}
\arguments{
\item{theData}{usually the result from \code{link{getData.All}}}

\item{closeRts}{list object with tolerances & parameters for peak finding.
See note with the function \code{link{getRtArea}} for a description of the
different items. Please note that in stead of retention times, for this
function it's actually relative retention times that are used here, so
tolerances/limits should be tighter.}

\item{refcloseRts}{list object with tolerances & parameters for peak finding
for the reference compound in the targets data.frame (marked by the column
Reference, where ONE compound is set to TRUE)
See note with the function \code{link{getRtArea}} for a description of the
different items. The items are the same as for the closeRts.}

\item{fixedRrt}{numeric vector (default is NA). If not NA, then this value is
used as the result of this function. This serves as a means to 'set' a
reference retention time when integration or finding the right peak is
not possible.}

\item{includeMissing}{default is TRUE. If false then an empty data-row will
be generated when nothing is found. If TRUE then the compound data is in the
data row with NA for all missing data}

\item{mzNr}{rownumber in the targets data.frame to search the peak info for.
If NA then all rows will be searched individually (may take some time) and
returned in data.frame. If integer value then it will search only for the
'component' in that specific row of the targets data.frame. If a component
is not found, them the peak info will be all NA's}
}
\value{
data.frame
}
\description{
Does the same thing as \code{link{getRtArea}}, but works with relative
 retention times. It first attempts to find the peak information on the
 reference compound in the tergets data.frame and then calculates the
 relative retention times with this. These are then used to search the peak
 info tables (with the retention times in there also 'converted' to relative
 retention times).
}
\note{
Through testing this function was found to be more dependable, because
 it allows some absolute shifts in Rt. If there are very big shifts, it may
 be necessary to increase 'left' and 'right' variables in the closeRts
 parameter.
}
\examples{
demoSpecFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005.mzdata.xml",
  package = "AgilentDataV2"
)
demoPeaksFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005_MS.csv",
  package = "AgilentDataV2"
)
demoTargetsFile <- fs::path_package(
  "extdata",
  "targets/targets001.csv",
  package = "AgilentDataV2"
)
result <- getData.All(
  sampleName = "WKL-0005",
  peptideTargetsFile = demoTargetsFile,
  mzDataFile = demoSpecFile,
  msPeakFile = demoPeaksFile,
)
getRRtArea(theData = result[[1]], mzNr = 1)
getRRtArea(theData = result[[1]])
}
