% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peakList.R
\name{getRRtReference}
\alias{getRRtReference}
\title{getRRtReference}
\usage{
getRRtReference(
  theData,
  closeRts = list(left = 0.5, right = 0.5, closest = TRUE, peak = FALSE),
  fixedRrt = NA
)
}
\arguments{
\item{theData}{usually the result from \code{link{getData.All}}}

\item{closeRts}{list object with tolerances & parameters for peak finding.
See note for description of the different items}

\item{fixedRrt}{numeric vector (default is NA). If not NA, then this value is
used as the result of this function. This serves as a means to 'set' a
reference retention time when integration or finding the right peak is
not possible.}
}
\value{
a numeric vector
}
\description{
Function to find the reference retention from a list-object containing
  both a targets table and a peakList table. This depends on the targets
  table having a single component being reference = TRUE
}
\note{
Items in the closeRts: left & right = tolerances for finding the peak
 retention time in the targets table. The function searches first for an EIC
 in the peak info data (files section of the msInfo object) that is the exact
 same (4 digits) as the one of the component to be searched for. Peaks that
 have a retention that falls in between rt - left and rt + right are
 considered. If the 'closest' variable is set to TRUE, the function will take
 the peak with the Rt closest to the Rt in the targets data.frame. If False,
 then if the 'peak' variable is TRUE, the it will sort the found peaks that
 fall within the tolerances (rt-left, rt+right) and take the nth peak as
 specfified by the 'Peak' column in the targets column. This can be needed to
 deal accurately with close together peaks in the same EIC. If the 'peak'
 variable is FALSE, then it will select the peak with the highest Area.
}
\examples{
demoSpecFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005.mzdata.xml",
  package = "AgilentDataV2"
)
demoPeaksFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005_MS.csv",
  package = "AgilentDataV2"
)
demoTargetsFile <- fs::path_package(
  "extdata",
  "targets/targets001.csv",
  package = "AgilentDataV2"
)
result <- getData.All(
  sampleName = "WKL-0005",
  peptideTargetsFile = demoTargetsFile,
  mzDataFile = demoSpecFile,
  msPeakFile = demoPeaksFile,
)
getRRtReference(theData = result)
getRRtReference(theData = result, fixedRrt = 10)
}
