% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peakList.R
\name{getRtArea}
\alias{getRtArea}
\title{getRtArea}
\usage{
getRtArea(
  theData,
  closeRts = list(left = 0.5, right = 0.5, closest = FALSE, peak = TRUE),
  includeMissing = TRUE,
  mzNr = NA
)
}
\arguments{
\item{theData}{usually the result from \code{link{getData.All}}}

\item{closeRts}{list object with tolerances & parameters for peak finding.
See note for description of the different items}

\item{includeMissing}{default is TRUE. If false then an empty data-row will
be generated when nothing is found. If TRUE then the compound data is in the
data row with NA for all missing data}

\item{mzNr}{rownumber in the targets data.frame to search the peak info for.
If NA then all rows will be searched individually (may take some time) and
returned in data.frame. If integer value then it will search only for the
'component' in that specific row of the targets data.frame. If a component
is not found, them the peak info will be all NA's}
}
\value{
data.frame
}
\description{
function that searches the peak info data for the peak closest
 to the Rt (retention time) specified in the targets data.frame. Data
 provided needs to contain at least the peak info data in the files section
 of the msInfo object of the the parameter 'theData' and the target info in
 the 'targets' data.frame of the 'theData' parameter
}
\note{
Items in the closeRts: left & right = tolerances for finding the peak
 retention time in the targets table. The function searches first for an EIC
 in the peak info data (files section of the msInfo object) that is the exact
 same (4 digits) as the one of the component to be searched for. Peaks that
 have a retention that falls in between rt - left and rt + right are
 considered. If the 'closest' variable is set to TRUE, the function will take
 the peak with the Rt closest to the Rt in the targets data.frame. If False,
 then if the 'peak' variable is TRUE, the it will sort the found peaks that
 fall within the tolerances (rt-left, rt+right) and take the nth peak as
 specfified by the 'Peak' column in the targets column. This can be needed to
 deal accurately with close together peaks in the same EIC. If the 'peak'
 variable is FALSE, then it will select the peak with the highest Area.

In most cases, it was found that closeRts = (..., ..., FALSE, TRUE)
 works best
}
\examples{
demoSpecFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005.mzdata.xml",
  package = "AgilentDataV2"
)
demoPeaksFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005_MS.csv",
  package = "AgilentDataV2"
)
demoTargetsFile <- fs::path_package(
  "extdata",
  "targets/targets001.csv",
  package = "AgilentDataV2"
)
result <- getData.All(
  sampleName = "WKL-0005",
  peptideTargetsFile = demoTargetsFile,
  mzDataFile = demoSpecFile,
  msPeakFile = demoPeaksFile,
)
getRtArea(theData = result[[1]], mzNr = 1)
getRtArea(theData = result[[1]])
}
