% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tools.R
\name{giveFilePathsFromStrings}
\alias{giveFilePathsFromStrings}
\title{giveFilePathsFromStrings}
\usage{
giveFilePathsFromStrings(strings)
}
\arguments{
\item{strings}{character vector, usually the result of
\code{link[base]{list.files}}}
}
\value{
character vector (can be character(0))
}
\description{
Function that extracts the directory (path) part from a set of file
 locations. This function eliminates character vectors parts that contain the
 substring '.mzdata' (files with extension '.mzdata.xml') and names that have
 the pattern '_.*$'
}
\examples{
giveFilePathsFromStrings("R:/testfiles/test.mzdata.xml")
giveFilePathsFromStrings(c("R:/testfiles/test.xml", "R:/test2/test5.csv"))
giveFilePathsFromStrings(c("R:/testfiles/test.xml", "R:/test2/test5.csv",
                           "R:/testfiles/test.mzdata.xml"))
giveFilePathsFromStrings(c("R:/testfiles/test.xml", "R:/test2/test5.csv",
                           "R:/testfiles/test6.mzdata.xml"))
}
