% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_saveSignals.R
\name{giveImpurity}
\alias{giveImpurity}
\title{giveImpurity}
\usage{
giveImpurity(theModification)
}
\arguments{
\item{theModification}{character vector}
}
\value{
if theModification is "DON100" then "B", if theModification is
 "DON82" then "A". If theModification is something else then "" is returned.
}
\description{
function that translates the "DON00" modification to "B" and the
 "DON82" modification to "A"
}
\examples{
giveImpurity("DON100")
giveImpurity("DON82")
giveImpurity("DON")
giveImpurity("Nothing")
}
