% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_saveSignals.R
\name{giveLabel}
\alias{giveLabel}
\title{giveLabel}
\usage{
giveLabel(theModification)
}
\arguments{
\item{theModification}{character vector describing the modification(s) on a
peptide}
}
\value{
character vector, either 'DON' or ''
}
\description{
function that correctly queries a modification for containing
 DON-label. This is needed for the more complicated situation where there are
 two labels on a peptide and the modification (which is a description of
 modification(s) present on a peptide) has two elements
}
\examples{
giveLabel("DON")
giveLabel("DON, DON")
giveLabel("DON82")
}
