% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_saveSignals.R
\name{giveLocation}
\alias{giveLocation}
\title{giveLocation}
\usage{
giveLocation(theSequence = "TQINKVVR")
}
\arguments{
\item{theSequence}{character vector representing the aminoacid sequence of a
peptide}
}
\value{
character vector (can be NA)
}
\description{
function to provide modification locations for a peptide
 sequence. This is needed as the 'native' peptide for a potential labelling
 location does not show a (potential) labelling location in the target table.
}
\examples{
giveLocation()
giveLocation("DILNQITKPNDVYSFSLASR")
giveLocation("VASMASEK")
}
