% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tools.R
\name{ifelseProper}
\alias{ifelseProper}
\title{ifelseProper}
\usage{
ifelseProper(logicValue = NULL, ifTrue = NULL, ifFalse = NULL)
}
\arguments{
\item{logicValue}{variable or expression resulting in TRUE or FALSE,
if missing or not logical then the function will return NULL.}

\item{ifTrue}{variable or expression to be returned when logicValue == TRUE}

\item{ifFalse}{variable or expression to be returned when logicValue == FALSE}
}
\value{
depending on logicValue, ifTrue or ifFalse
}
\description{
\code{link[base]{basename}} replacement for properly returning all
 datatypes. Note: unlike the regular function, this one is NOT vectorized.
 Also: no optimization, so there will be a performance 'penalty'
}
\examples{
ifelse(mtcars$cyl == 6, "Six", "Other")
ifelse(mtcars$cyl == 6, list("Six", TRUE), list("Other", FALSE))
ifelse(mtcars$cyl == 6, list("Six", TRUE), list("Other", FALSE))[[1]]
ifelse(mtcars$cyl[1] == 6, list("Six", TRUE), list("Other", FALSE))
ifelseProper(mtcars$cyl[1] == 6, list("Six", TRUE), list("Other", FALSE))
purrr::map(mtcars$cyl, ~ifelseProper(.x == 6, list("Six", TRUE), list("Other", FALSE)))
purrr::map(mtcars$cyl, ~ifelseProper(.x == 6, list("Six", TRUE), list("Other", FALSE)))[[1]]
}
