% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectraAgilent.R
\name{mzData}
\alias{mzData}
\title{mzData}
\value{
a new 'mzData' object

data.frame, with columns 'mz' and 'intensity', or NA

data.frame or NA

dataElement object
}
\description{
R6 Class dealing with mass spectrum data as output by Agilent Masshunter.
  Main drawback of this class is that currently all data needs to be loaded
  into memory. The .mzData.xml files can be quite big, especially when using
  high quality profile data. Profile data may also give issues with peak
  detection as peaks on some equipment have a tendency to be assymmetric.
  Consider using centroided data and/or higher background cutoff when
  acquiring
}
\note{
This may take a while since data needs to be decoded and organized
 into a proper data.frame
}
\examples{
demoFile <- fs::path_package(
  "extdata",
  "raw/WKL-0001.mzdata.xml",
  package = "AgilentDataV2"
)
demoFile
spectrData <- mzData$new(filename = demoFile)
spectrData
spectrData$info(centroided = TRUE)
plot(
 spectrData$spectrum(number = 1),
 type = "h",
 xlab = "m/z", ylab = 'Intensity'
)
plot(
 spectrData$spectrum(number = 2),
 type = "h",
 xlab = "m/z", ylab = 'Intensity'
)
demoFile <- fs::path_package(
  "extdata",
  "raw/WorklistData-0005.mzdata.xml",
  package = "AgilentDataV2"
)
demoFile
spectrData <- mzData$new(filename = demoFile)
spectrData
spectrData$info(centroided = TRUE)
plot(
 spectrData$spectrum(number = 3),
 type = "h",
 xlab = "m/z", ylab = 'Intensity'
)
plot(
 spectrData$spectrum(number = 4),
 type = "h",
 xlab = "m/z", ylab = 'Intensity'
)

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{empty}}{if TRUE, then data has been read. If FALSE, then not.
Read only}

\item{\code{filename}}{shows filename after reading data into the object
Read only}

\item{\code{endian}}{gets/sets the type of endian ('little' or 'big') to be used
when dealing with the .mzdata.xml data}

\item{\code{data}}{provides access to the 'raw' data after it has been read from
the file.}

\item{\code{length}}{displays the number of spectra present in the object.
Read only}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mzData-new}{\code{mzData$new()}}
\item \href{#method-mzData-print}{\code{mzData$print()}}
\item \href{#method-mzData-spectrum}{\code{mzData$spectrum()}}
\item \href{#method-mzData-info}{\code{mzData$info()}}
\item \href{#method-mzData-read}{\code{mzData$read()}}
\item \href{#method-mzData-clone}{\code{mzData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mzData-new"></a>}}
\if{latex}{\out{\hypertarget{method-mzData-new}{}}}
\subsection{Method \code{new()}}{
initializes the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mzData$new(filename = as.character(NA), endian = "little", message = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of the .mzData.xml file (exported by Agilent Software)}

\item{\code{endian}}{specifies whether the base64 encoded data in the .mzData.xml
file is 'little' or 'big' endian}

\item{\code{message}}{logical vector. If TRUE (default) a message will be printed
when the data has been loaded. Files are sometimes big and may take a while
to load}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mzData-print"></a>}}
\if{latex}{\out{\hypertarget{method-mzData-print}{}}}
\subsection{Method \code{print()}}{
For printing purposes: prints info on the data loaded into the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mzData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mzData-spectrum"></a>}}
\if{latex}{\out{\hypertarget{method-mzData-spectrum}{}}}
\subsection{Method \code{spectrum()}}{
retrieves a spectrum from the (internal) .mzData.xml data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mzData$spectrum(number = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{number}}{number of spectrum to retrieve}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mzData-info"></a>}}
\if{latex}{\out{\hypertarget{method-mzData-info}{}}}
\subsection{Method \code{info()}}{
returns a data.frame with info on spectra in the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mzData$info(number = NA, centroided = FALSE, setSingleRt = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{number}}{integer vector. One or more numbers of the spectra for
which the info is to be retrieved}

\item{\code{centroided}}{logical vector. Adds the centroided column to the
info data.frame. Sometimes needed for plotting and analysis purposes.}

\item{\code{setSingleRt}}{numeric vector: retention time in minutes to subtract and
add to the retention time of single spectrum spectra. Spectra are usually a
sum or average of a number of spectra close together. The lowest Rt spectrum
provides the start (Rt), the highest Rt spectrum provides the end (Rt). In
case a single spectrum provides the data then start = Rt - setSingleRt and
end = Rt + setSingleRt. This is to accomodate other code where both start
and end are required}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mzData-read"></a>}}
\if{latex}{\out{\hypertarget{method-mzData-read}{}}}
\subsection{Method \code{read()}}{
function factory that generates a function to get a spectrum
 in the object and to return it as a \link[dataInfo]{dataElement} (for use with
 \link[MS.Analysis]{msInfo})
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mzData$read(number = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{number}}{integer value (single value!) specifying which spectrum to
retrieve from the object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mzData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mzData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mzData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
