% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readAgilent.R
\name{peakListInfo.AgilentExport}
\alias{peakListInfo.AgilentExport}
\title{peakListInfo.AgilentExport}
\usage{
peakListInfo.AgilentExport(
  string,
  defaultCollapse = ";",
  translateComment = chromatogramInfo.AgilentExport
)
}
\arguments{
\item{string}{character vector to be translated}

\item{defaultCollapse}{only used in case of unknown traces. For these, the
resulting elements are joined together into a single character vector which
can be split by this character}

\item{translateComment}{defines the function to be used to translate the
description line. Default is 'chromatogramInfo.AgilentExport2'.}
}
\value{
a data.frame
}
\description{
Replacement function for \code{link[readAgilent]{readPeaklist.AgilentExport}}
 Has an additional argument to allow for different 'translating' of the
 description line in peaklist export files
}
\examples{
info <- "D:/MassHunter/Data/Ben/Data0001.d, +ESI TIC Scan Frag=125.0V Data0001.d "
peakListInfo.AgilentExport(info)
info <- "D:/MassHunter/Data/Ben/Data0001.d, +ESI EIC(591.2807) Scan Frag=125.0V Data0001.d "
peakListInfo.AgilentExport(info)

}
