% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{processingFilesEmpty}
\alias{processingFilesEmpty}
\title{processingFilesEmty}
\usage{
processingFilesEmpty()
}
\value{
a two element list of chracter vectors with the names 'MSPeakFiles'
 and 'Spectra'
}
\description{
Helper function function that gives an empty files list
}
\examples{
processingFilesEmpty()
}
