% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tools.R
\name{readFile}
\alias{readFile}
\title{readFile}
\usage{
readFile(
  filename,
  readFunction = XML::xmlToList,
  stopOnError = FALSE,
  showWarning = TRUE
)
}
\arguments{
\item{filename}{name of the file (including path) to be read. Note: filenames
are automatically trimmed (removal of spaces).}

\item{readFunction}{specifies the function to used to read the file}

\item{stopOnError}{logical vector specifying if execution should stop on
error. Default is FALSE}

\item{showWarning}{logical vector specifying whether to show warnings.
Default is TRUE}
}
\value{
a function for reading the file with name filename (incl path).
 This function will return NA if reading the file fails. Execution will not
 stop unless specified.
}
\description{
function factory that generates a file reading function that can take into
 account that a reading file action may fail. It allows for use of different
 read functions to be used via functional programming. The result of the
 function will be NA if the file read action fails.
}
\examples{
filename <- tempfile()
write.csv(mtcars, filename)
readFile(filename = filename, readFunction = read.csv)()
write.table(mtcars, filename, col.names = TRUE, sep = ";")
readFile(filename = filename, readFunction = read.csv)(header = TRUE, sep = ";")
file.remove(filename)
readFile(filename = filename, readFunction = read.csv)(header = TRUE, sep = ";")
}
