% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readAgilent.R
\name{readPeaklist.AgilentExport.memory}
\alias{readPeaklist.AgilentExport.memory}
\title{readPeaklist.AgilentExport.memory}
\usage{
readPeaklist.AgilentExport.memory(
  textLines,
  sep = ",",
  startsString,
  useSmooth = FALSE,
  translateComment = chromatogramInfo.AgilentExport
)
}
\arguments{
\item{textLines}{character vector of the data in Agilent peaklist export
format: first line = description and the rest of the lines is the integration
data. Rownumbers are ignored.}

\item{sep}{defines the separator for the peaklist data. Default is ','}

\item{startsString}{character vector that defines which lines are the start
of a peaklist lines in textLines}

\item{useSmooth}{logical vector that defines whether smoothed data is removed
(FALSE) or if raw data is removed (TRUE). If NA, then nothing, then no data
is removed. Note: When chromatograms are smoothed, we end up with two
chromatograms per trace: the raw & the smoothed one. If smoothing is done
prior to peak detection/integration, then there will also be two peak tables
per trace. This can lead to confusion in code where data is combined. This
option removes either the smoothed data or the raw data. Default is FALSE}

\item{translateComment}{defines the function to be used to translate the
description line. Default is 'chromatogramInfo.AgilentExport'.}
}
\value{
function that generates a list of two objects: first element (info)
 is data.frame (with info) and second (data) is a list of data.frame's
}
\description{
replacement function (factory) for \code{link[readAgilent]{readPeaklist.AgilentExport.memory}}
 Original implementation did not allow for custom translateComment functions. This one does.
}
\note{
a peaklist = the result from the integration of traces/chromatograms by
 the Agilent Masshunter software
}
\examples{
demoFile <- fs::path_package("extdata", "raw/WorklistData-0005_MS.csv", package = "AgilentDataV2")
result <- readLines(demoFile)
result <- readPeaklist.AgilentExport.memory(result, startsString = "WorklistData-0005")()
result$info
result$data[[1]]
result$data[[3]]
}
