% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readAgilent.R
\name{readPeaklist.AgilentExport}
\alias{readPeaklist.AgilentExport}
\title{readPeaklist.AgilentExport}
\usage{
readPeaklist.AgilentExport(
  filename,
  sep = ",",
  startsString =
    paste0(stringr::str_remove(tools::file_path_sans_ext(basename(filename)), pattern =
    "_.*"), ".d"),
  useSmooth = FALSE,
  translateComment = chromatogramInfo.AgilentExport
)
}
\arguments{
\item{filename}{name of the peaklist file}

\item{sep}{defines the separator for the peaklist data. Default is ','}

\item{startsString}{character vector that defines which lines are the start
of a peaklist lines in the file}

\item{useSmooth}{logical vector that defines whether smoothed data is removed
(FALSE) or if raw data is removed (TRUE). If NA, then nothing, then no data
is removed. Note: When chromatograms are smoothed, we end up with two
chromatograms per trace: the raw & the smoothed one. If smoothing is done
prior to peak detection/integration, then there will also be two peak tables
per trace. This can lead to confusion in code where data is combined. This
option removes either the smoothed data or the raw data. Default is FALSE}

\item{translateComment}{defines the function to be used to translate the
description line. Default is 'chromatogramInfo.AgilentExport'}
}
\value{
function that generates a list of two objects: first element (info)
 is data.frame (with info) and second (data) is a list of data.frame's or NA
 (when a reading from a file fails)
}
\description{
replacement function (factory) for \code{link[readAgilent]{readPeaklist.AgilentExport}}
 Uses 'readFile' function to allow for failed read reads of the file. Note:
 will return NA, when reading of the file fails.
}
\note{
a peaklist = the result from the integration of traces/chromatograms by
 the Agilent Masshunter software
}
\examples{
demoFile <- fs::path_package("extdata", "raw/WorklistData-0005_MS.csv", package = "AgilentDataV2")
result <- readPeaklist.AgilentExport(demoFile, startsString = "WorklistData-0005")()
length(result)
result[[1]]
result$info
result$data[[1]]
result$data[[4]]

}
