% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tools.R
\name{selectStrings}
\alias{selectStrings}
\title{selectStrings}
\usage{
selectStrings(
  strings,
  positive = NA,
  negative = NA,
  returnStrings = TRUE,
  useBasename = TRUE,
  stopOnError = FALSE,
  showWarning = TRUE
)
}
\arguments{
\item{strings}{character vector to be filtered}

\item{positive}{character vector to be used as a positive criterium pattern
via \code{link[base]{grepl}}. To put it simple if the 'positive' character
is present in the 'strings' character vector then it will be included, see
examples}

\item{negative}{character vector to be used as a negative criterium pattern
via \code{link[base]{grepl}}. To put it simple if the 'negative' character
is present in the 'strings' character vector then it will not be included,
see examples}

\item{returnStrings}{logical vector determines whether to return the actual
strings (which comply with the positive and the negative criteria) (default)
or the position/ See examples}

\item{useBasename}{logical vector. If TRUE, then the filtering takes place
after \code{link[base]{basename}} has been appied on the 'strings' argument.
Note that the returned strings (when returnStrings is TRUE) will be the
original values.}

\item{stopOnError}{logical vector specifying if execution should stop on
error. Default is FALSE}

\item{showWarning}{logical vector specifying whether to show warnings.
Default is TRUE}
}
\value{
character or integer vector
}
\description{
select a subset from a character vector base on a substring that should be
 present (positive criterium) and/or a substring that should not be present
 (negative criterium). Originally meant to be able to filter filenames, but
 can be used on other character vectors
}
\examples{
selectStrings(strings = c("test_MS.csv", "testMS.csv"),
              positive = "_MS.csv$")
selectStrings(strings = c("test_MS.csv", "testMS.csv"),
              positive = "MS.csv$")
selectStrings(strings = c("test_MS.csv", "testMS.csv"),
              positive = "MS.csv$", negative = "_")
selectStrings(strings = c("test_MS.csv", "testMS.csv"),
              positive = "MS.csv$", negative = "_",
              returnStrings = FALSE)
selectStrings(strings = c("test_MS.csv", "testMS.csv"),
              positive = "MS.csb$", negative = "_")
}
