% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{testCompleteData}
\alias{testCompleteData}
\title{testCompleteData}
\usage{
testCompleteData(theFilenames)
}
\arguments{
\item{theFilenames}{two element named list of character vectors}
}
\value{
named logical vector
}
\description{
Function that takes a set of filenames and attempts to determine whether
 for every MSPeakfiles there is a spectrum file (and vice versa)
}
\examples{
processingFiles <- list(
  MSPeakFiles = c("./20250702/WorklistData-0003_MS.csv",
                  "./20250702/WorklistData-0005_MS.csv"),
  Spectra = c("./20250702/WorklistData-0003.mzdata.xml",
              "./20250702/WorklistData-0005.mzdata.xml")
)
testCompleteData(processingFiles)
processingFilesError1 <- list(
  MSPeakFiles = c("./20250702/WorklistData-0003_MS.csv"),
  Spectra = c("./20250702/20250702/WorklistData-0003.mzdata.xml",
              "./20250702/WorklistData-0005.mzdata.xml")
)
testCompleteData(processingFilesError1)
processingFilesError2 <- list(
  MSPeakFiles = c("./20250702/WorklistData-0003_MS.csv",
                  "./20250702/WorklistData-0005_MS.csv"),
  Spectra = c("./20250702/WorklistData-0005.mzdata.xml")
)
testCompleteData(processingFilesError2)
}
